refactored some code into new ClearTransactionSelector component
