+++ src/components/TransactionsTab/TransactionsTab.styles.tsx
@@ -18,6 +18,7 @@ import {
   InfoHeading,
   FormLabel,
 } from "../Typography/Typography";
+import { StyledTooltip } from "./subcomponents/ClearTransactionSelector/ClearTransactionSelector.styles";
 import WalletInfoButton from "./subcomponents/WalletInfoButton/WalletInfoButton";
 import WalletMobileMenu from "./subcomponents/WalletMobileMenu/WalletMobileMenu";
 
@@ -114,18 +115,6 @@ export const LegendLine = styled.span`
   background: transparent;
 `;
 
-export const StyledTooltip = styled(Tooltip) <{
-  $isSelectorOpen: boolean;
-  $isTooltip: boolean
-}>`
-  display: ${({ $isSelectorOpen, $isTooltip }) => (!$isSelectorOpen && $isTooltip ? "flex" : "none")};
-  position: relative;
-  z-index: 3;
-  max-width: min-content;
-  margin-top: -2rem;
-  top: 1.25rem;
-  margin-left: 15rem;
-`;
 
 export const IconBinContainer = styled.a`
   display: flex;
@@ -210,25 +199,6 @@ export const IconContainer = styled.div`
   background-color: rgb(110, 118, 134, 0.1);
   color: ${(props) => props.theme.colors.lightGrey};
   border-radius: 50%;
-`;
-
-export const SelectWrapper = styled.div<{ $isOpen: boolean }>`
-  display: ${({ $isOpen }) => ($isOpen ? "flex" : "none")};
-  justify-content: end;
-  position: relative;
-  margin-top: -2rem;
-  top: 3.5rem;
-  right: 1.5rem;
-  height: 2rem;
-  color: ${({ theme }) =>
-    theme.name === "dark" ? theme.colors.white : theme.colors.primary};
-`;
-
-export const StyledDropdown = styled(Dropdown)`
-  ${SelectButtonText}
-
-  text-align: left;
-  width: 5rem;
 `;
 
 export const BackButton = styled(motion.button)`
